requirements.txt
jsonschema
numpy
openai
python-dotenv
